requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai>=1.0.0,<2.0.0
nltk==3.8.1
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
requests==2.31.0
python-dotenv==1.0.0
